requirements.txt
tensorflow
numpy
opencv-python
matplotlib
Pillow
